+++ src/components/pitches.rs
@@ -1,27 +1,17 @@
+use crate::components::pitch_event::PitchEvent;
 use crate::components::strikezone::{DEFAULT_SZ_BOT, DEFAULT_SZ_TOP};
 use crate::components::util::convert_color;
-
 use mlb_api::live::LiveResponse;
 use mlb_api::plays::PlayEvent;
 use tui::{
     style::{Color, Style},
     text::{Line, Span},
-    widgets::ListItem,
     widgets::canvas::Rectangle,
 };
 
-/// Used to display the pitch number next to the pitch type in the Canvas. Hopefully no one has at
-/// bat longer than 21 pitches.
-/// There have been eight at bats that were 18 pitches or more since 1988, with the longest at 21.
-pub const PITCH_IDX: &[&str] = &[
-    "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16",
-    "17", "18", "19", "20", "21",
-];
-pub const DEFAULT_IDX: &str = "-";
-
 #[derive(Debug, Default)]
 pub struct Pitches {
-    pub pitches: Vec<Pitch>,
+    pub pitch_events: Vec<PitchEvent>,
 }
 
 #[derive(Debug)]
@@ -36,6 +26,7 @@ pub struct Pitch {
     pub speed: f64,
     pub strike_zone_bot: f64,
     pub strike_zone_top: f64,
+    pub count: Count,
 }
 
 impl Default for Pitch {
@@ -50,12 +41,28 @@ impl Default for Pitch {
             speed: 0.0,
             strike_zone_bot: DEFAULT_SZ_BOT,
             strike_zone_top: DEFAULT_SZ_TOP,
+            count: Count::default(),
         }
     }
 }
 
-impl Pitch {
-    pub fn from_play(play: &PlayEvent) -> Self {
+#[derive(Clone, Debug, Default)]
+pub struct Count {
+    pub balls: u8,
+    pub strikes: u8,
+}
+
+impl From<mlb_api::plays::Count> for Count {
+    fn from(value: mlb_api::plays::Count) -> Self {
+        Self {
+            balls: value.balls,
+            strikes: value.strikes,
+        }
+    }
+}
+
+impl From<&PlayEvent> for Pitch {
+    fn from(play: &PlayEvent) -> Self {
         let pitch_data = match play.pitch_data.as_ref() {
             Some(p) => p,
             None => return Pitch::default(),
@@ -88,9 +95,12 @@ impl Pitch {
             index: play.pitch_number.unwrap_or_default(),
             strike_zone_bot: pitch_data.strike_zone_bottom.unwrap_or(DEFAULT_SZ_BOT),
             strike_zone_top: pitch_data.strike_zone_top.unwrap_or(DEFAULT_SZ_TOP),
+            count: play.count.clone().into(),
         }
     }
+}
 
+impl Pitch {
     /// Convert a pitch into a TUI Rectangle so it can be displayed in a Canvas.
     pub fn as_rectangle(&self) -> Rectangle {
         const SCALE: f64 = 12.0; // feet to inches
@@ -104,20 +114,19 @@ impl Pitch {
         }
     }
 
-    /// Convert a pitch into a TUI List item, displaying the pitch index, result (ball, strike, ect)
-    /// and pitch type (cutter, changeup, ect). For example:
-    /// "1  Foul | Four-Seam Fastball"
-    pub fn as_list_item(&self, debug: bool) -> ListItem {
-        ListItem::new(vec![Line::from(vec![
+    /// Convert a pitch into a TUI Line item, displaying the pitch index, result (ball, strike, ect)
+    /// and pitch type (cutter, changeup, ect). For example: "1  Foul | Four-Seam Fastball"
+    pub fn as_lines(&self, debug: bool) -> Vec<Line> {
+        vec![Line::from(vec![
             Span::styled(format!(" {} ", self.index), Style::default().fg(self.color)),
             Span::raw(self.format(debug)),
-        ])])
+        ])]
     }
 
     fn format(&self, debug: bool) -> String {
         let s = format!(
-            " {:<20}| {:^5.1}| {}",
-            self.description, self.speed, self.pitch_type
+            " {:<20}| {}-{} | {:^5.1}| {}",
+            self.description, self.count.balls, self.count.strikes, self.speed, self.pitch_type
         );
         if debug {
             return format!(" {} | {:?}", s, self.location);
@@ -127,25 +136,16 @@ impl Pitch {
 }
 
 impl Pitches {
-    pub fn new(pitches: Vec<Pitch>) -> Self {
-        Pitches { pitches }
-    }
-
     pub fn from_live_data(live_game: &LiveResponse) -> Self {
-        let pitch_data = match live_game.live_data.plays.current_play.as_ref() {
-            Some(c) => Pitches::transform_pitches(&c.play_events),
+        let pitch_events = match live_game.live_data.plays.current_play.as_ref() {
+            Some(c) => Pitches::transform_pitch_events(&c.play_events),
             None => return Pitches::default(),
         };
-        Pitches::new(pitch_data)
+        Pitches { pitch_events }
     }
 
-    fn transform_pitches(plays: &[PlayEvent]) -> Vec<Pitch> {
-        plays
-            .iter()
-            .filter(|play| play.is_pitch)
-            .map(Pitch::from_play)
-            .rev()
-            .collect()
+    fn transform_pitch_events(plays: &[PlayEvent]) -> Vec<PitchEvent> {
+        plays.iter().map(PitchEvent::from).rev().collect()
     }
 }
 
@@ -153,6 +153,6 @@ impl Pitches {
 fn test_pitches_with_defaults() {
     // Testing what happens if there is no pitch data
     let play_event = vec![PlayEvent::default()];
-    let pitches = Pitches::transform_pitches(&play_event);
-    assert_eq!(pitches.len(), 0);
+    let pitches = Pitches::transform_pitch_events(&play_event);
+    assert_eq!(pitches.len(), 1);
 }